Se hizo una refactorizacion al controlador de usuarios y se creo un user.services para mantener alli la logica
